# MyRVM-Server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRVM API Documentation",
        "version": "1.0.0",
        "description": "API documentation for MyReverseVendingMachine Server",
        "contact": {
            "email": "admin@myrvm.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/rvm-machines": {
            "get": {
                "tags": ["RVM Machines"],
                "summary": "List all RVM machines",
                "operationId": "getRvmMachines",
                "security": [{"bearerAuth": []}],
                "responses": {
                    "200": { "description": "List of machines" },
                    "401": { "description": "Unauthorized" },
                    "500": { "description": "Internal Server Error" }
                }
            },
            "post": {
                "tags": ["RVM Machines"],
                "summary": "Create new RVM machine",
                "operationId": "createRvmMachine",
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["name","location","serial_number"],
                                "properties": {
                                    "name": { "type": "string", "example": "RVM Station 1" },
                                    "location": { "type": "string", "example": "Mall Central Park" },
                                    "serial_number": { "type": "string", "example": "RVM-JKT-001" },
                                    "status": { "type": "string", "enum": ["online","offline","maintenance"], "example": "offline" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Machine created successfully" },
                    "400": { "description": "Bad Request" },
                    "401": { "description": "Unauthorized" },
                    "422": { "description": "Unprocessable Entity - Validation Error" },
                    "500": { "description": "Internal Server Error" }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}